Renamed Wearable to Item and refactored HealthPotion to inherit from it.
