+++ src/main/java/entity/item/wearable/weapon/Weapon.java
@@ -1,9 +1,9 @@
 package entity.item.wearable.weapon;
 
 import entity.Creature;
-import entity.item.wearable.Wearable;
+import entity.item.Item;
 
-public class Weapon extends Wearable {
+public class Weapon extends Item {
 
     public Weapon(int damageBonus) {
         super(damageBonus);
